+++ packages/bottom-sheet/src/component.tsx
@@ -37,11 +37,18 @@ import {
     TIMEOUT,
 } from './utils';
 
+import defaultColors from './default.module.css';
 import styles from './index.module.css';
+import invertedColors from './inverted.module.css';
 
 const { isNil } = fnUtils;
 
 const adjustContainerHeightDefault = (value: number) => value;
+
+const colorStyles = {
+    default: defaultColors,
+    inverted: invertedColors,
+} as const;
 
 export const BottomSheet = forwardRef<HTMLDivElement, BottomSheetProps>(
     // eslint-disable-next-line complexity
@@ -55,17 +62,21 @@ export const BottomSheet = forwardRef<HTMLDivElement, BottomSheetProps>(
             titleSize = 'default',
             subtitle,
             actionButton,
+            bottomSheetWrapperClassName,
             contentClassName,
             containerClassName,
             containerProps,
             headerClassName,
+            headerContentClassName,
             footerClassName,
             addonClassName,
+            outerClassName,
             closerClassName,
             backerClassName,
             modalClassName,
             modalWrapperClassName,
             className,
+            outerAddons,
             leftAddons,
             rightAddons,
             bottomAddons,
@@ -81,6 +92,7 @@ export const BottomSheet = forwardRef<HTMLDivElement, BottomSheetProps>(
             hideHeader,
             disableOverlayClick,
             disableBlockingScroll,
+            scrollLock = true,
             disableFocusLock,
             children,
             zIndex,
@@ -115,6 +127,8 @@ export const BottomSheet = forwardRef<HTMLDivElement, BottomSheetProps>(
             backButtonProps,
             iOSLock = false,
             virtualKeyboard = false,
+            colors = 'default',
+            preventScrollOnSwipe,
         },
         ref,
     ) => {
@@ -167,6 +181,8 @@ export const BottomSheet = forwardRef<HTMLDivElement, BottomSheetProps>(
         const emptyHeader = !hasCloser && !leftAddons && !title && !hasBacker && !rightAddons;
         const titleIsReactElement = React.isValidElement(title);
 
+        const colorStyle = colorStyles[colors];
+
         const headerProps: HeaderProps = {
             ...(titleIsReactElement
                 ? { children: title }
@@ -174,7 +190,9 @@ export const BottomSheet = forwardRef<HTMLDivElement, BottomSheetProps>(
             scrollableParentRef: scrollableContainer,
             headerRef,
             headerOffset,
-            className: headerClassName,
+            colors,
+            className: cn(headerClassName, colorStyle.hasContent),
+            contentClassName: headerContentClassName,
             addonClassName,
             closerClassName,
             backButtonClassName: backerClassName,
@@ -483,6 +501,7 @@ export const BottomSheet = forwardRef<HTMLDivElement, BottomSheetProps>(
             trackMouse: swipeable,
             trackTouch: swipeable,
             delta: swipeThreshold,
+            preventScrollOnSwipe,
         });
 
         const handleExited = (node: HTMLElement) => {
@@ -586,6 +605,7 @@ export const BottomSheet = forwardRef<HTMLDivElement, BottomSheetProps>(
                         className={cn(
                             styles.marker,
                             styles.defaultMarker,
+                            colorStyle.defaultMarker,
                             swipeableMarkerClassName,
                         )}
                     />
@@ -595,7 +615,9 @@ export const BottomSheet = forwardRef<HTMLDivElement, BottomSheetProps>(
             return null;
         };
 
-        const bgClassName = backgroundColor && styles[`background-${backgroundColor}`];
+        const bgClassName =
+            backgroundColor &&
+            styles[`background-${backgroundColor}${colors === 'inverted' ? '-inverted' : ''}`];
 
         return (
             <BaseModal
@@ -637,18 +659,33 @@ export const BottomSheet = forwardRef<HTMLDivElement, BottomSheetProps>(
                 disableRestoreFocus={disableRestoreFocus}
                 keepMounted={keepMounted}
                 iOSLock={iOSLock}
+                scrollLock={scrollLock}
             >
                 <div
-                    className={cn(styles.wrapper, {
+                    className={cn(styles.wrapper, bottomSheetWrapperClassName, {
                         [styles.fullscreen]: headerOffset === 0 && sheetOffset === 0,
                     })}
                     onTransitionEnd={setSheetHeight}
                 >
+                    {outerAddons && (
+                        <div
+                            className={cn(styles.outerClassName, outerClassName)}
+                            style={getSwipeStyles()}
+                        >
+                            {outerAddons}
+                        </div>
+                    )}
                     <div
-                        className={cn(styles.component, bgClassName, className, {
-                            [styles.withTransition]: swipingInProgress === false,
-                            [styles.safeAreaBottom]: os.isIOS(),
-                        })}
+                        className={cn(
+                            styles.component,
+                            bgClassName,
+                            colorStyle.component,
+                            className,
+                            {
+                                [styles.withTransition]: swipingInProgress === false,
+                                [styles.safeAreaBottom]: os.isIOS(),
+                            },
+                        )}
                         style={{
                             ...getSwipeStyles(),
                             ...getHeightStyles(),
@@ -677,10 +714,15 @@ export const BottomSheet = forwardRef<HTMLDivElement, BottomSheetProps>(
 
                             <div
                                 ref={contentRef}
-                                className={cn(styles.content, contentClassName, {
-                                    [styles.noHeader]: hideHeader || emptyHeader,
-                                    [styles.noFooter]: !actionButton,
-                                })}
+                                className={cn(
+                                    styles.content,
+                                    colorStyle.content,
+                                    contentClassName,
+                                    {
+                                        [styles.noHeader]: hideHeader || emptyHeader,
+                                        [styles.noFooter]: !actionButton,
+                                    },
+                                )}
                                 data-test-id={getDataTestId(dataTestId, 'content')}
                             >
                                 {children}
@@ -690,6 +732,7 @@ export const BottomSheet = forwardRef<HTMLDivElement, BottomSheetProps>(
                                 <Footer
                                     sticky={stickyFooter}
                                     className={cn(bgClassName, footerClassName)}
+                                    colors={colors}
                                     dataTestId={getDataTestId(dataTestId, 'footer')}
                                 >
                                     {actionButton}