+++ packages/bottom-sheet/src/types.ts
@@ -1,10 +1,13 @@
 import { type HTMLAttributes, type ReactElement, type ReactNode, type RefObject } from 'react';
+import { type SwipeableProps } from 'react-swipeable';
 import { type HandledEvents } from 'react-swipeable/es/types';
 import { type TransitionProps } from 'react-transition-group/Transition';
 
 import { type BaseModalProps } from '@alfalab/core-components-base-modal';
 import { type NavigationBarPrivateProps as NavigationBarProps } from '@alfalab/core-components-navigation-bar-private';
 import { type BackgroundColorType } from '@alfalab/core-components-types';
+
+export type ColorType = 'default' | 'inverted';
 
 export type BottomSheetTitleAlign = 'center' | 'left';
 
@@ -61,6 +64,11 @@ export type BottomSheetProps = {
     className?: string;
 
     /**
+     * Дополнительный класс для обертки контейнера
+     */
+    bottomSheetWrapperClassName?: string;
+
+    /**
      * Дополнительный класс
      */
     contentClassName?: string;
@@ -84,6 +92,11 @@ export type BottomSheetProps = {
      * Дополнительный класс шапки
      */
     headerClassName?: string;
+
+    /**
+     * Дополнительный класс для контента шапки
+     */
+    headerContentClassName?: string;
 
     /**
      * Дополнительный класс футера
@@ -164,6 +177,16 @@ export type BottomSheetProps = {
     swipeThreshold?: number;
 
     /**
+     * Дополнительный класс для слота над шторкой
+     */
+    outerClassName?: string;
+
+    /**
+     * Слот над шторкой
+     */
+    outerAddons?: ReactNode;
+
+    /**
      * Слот слева
      */
     leftAddons?: ReactNode;
@@ -235,8 +258,15 @@ export type BottomSheetProps = {
 
     /**
      * Отключает блокировку скролла при открытии модального окна
+     * @deprecated Используйте `scrollLock={true}`.
      */
     disableBlockingScroll?: boolean;
+
+    /**
+     * Управляет блокировкой скролла/overscroll фона при открытой шторке.
+     * @default true
+     */
+    scrollLock?: BaseModalProps['scrollLock'];
 
     /**
      * Отключает ловушку фокуса
@@ -354,6 +384,7 @@ export type BottomSheetProps = {
 
     /**
      * Блокирует скролл когда модальное окно открыто. Работает только на iOS
+     * @deprecated Используйте `scrollLock={true}`.
      */
     iOSLock?: boolean;
 
@@ -361,4 +392,16 @@ export type BottomSheetProps = {
      * Учитывать высоту виртуальной клавиатуры
      */
     virtualKeyboard?: boolean;
+
+    /**
+     * Набор цветов для компонента
+     * @default default
+     */
+    colors?: 'default' | 'inverted';
+
+    /**
+     * Предотвращать скролл во время свайпа
+     * @default false
+     */
+    preventScrollOnSwipe?: SwipeableProps['preventScrollOnSwipe'];
 };