+++ packages/confirmation/src/components/base-confirmation/component.tsx
@@ -31,6 +31,7 @@ export const BaseConfirmation: FC<ConfirmationProps> = ({
     blockSmsRetry,
     dataTestId,
     className,
+    maxWidth,
     getScreensMap,
     onInputFinished,
     onChangeState,
@@ -135,6 +136,7 @@ export const BaseConfirmation: FC<ConfirmationProps> = ({
                 className={cn(styles.component, className, {
                     [styles.center]: alignContent === 'center',
                 })}
+                style={{ maxWidth }}
                 data-test-id={dataTestId}
             >
                 {CurrentScreen && <CurrentScreen mobile={mobile} />}