+++ packages/slider-input/src/Component.tsx
@@ -86,7 +86,7 @@ export type SliderInputProps = Omit<
     /**
      * Компонент поля ввода
      */
-    Input?: ComponentType<Omit<InputProps, 'onChange' | 'value'>>;
+    Input?: ComponentType<Omit<InputProps, 'onChange' | 'value' | 'defaultValue'>>;
 
     /**
      * Кастомные пропсы для поля ввода
@@ -137,6 +137,12 @@ export type SliderInputProps = Omit<
      * Идентификатор для систем автоматизированного тестирования
      */
     dataTestId?: string;
+
+    /**
+     * Жирность текста
+     * @default true
+     */
+    bold?: boolean;
 };
 
 const SIZE_TO_CLASSNAME_MAP = {
@@ -185,6 +191,7 @@ export const SliderInput = forwardRef<HTMLInputElement, SliderInputProps>(
             pips,
             range,
             dataTestId,
+            bold = true,
             ...restProps
         },
         ref,
@@ -258,6 +265,7 @@ export const SliderInput = forwardRef<HTMLInputElement, SliderInputProps>(
                         [styles.filled]: Boolean(value),
                         [styles.hasLabel]: label,
                         [styles.hasError]: Boolean(error),
+                        [styles.bold]: bold,
                     },
                     styles[SIZE_TO_CLASSNAME_MAP[size]],
                     className,
@@ -266,7 +274,6 @@ export const SliderInput = forwardRef<HTMLInputElement, SliderInputProps>(
             >
                 <Input
                     {...restProps}
-                    {...customInputProps}
                     ref={ref}
                     value={value.toString()}
                     onChange={handleInputChange}
@@ -320,6 +327,7 @@ export const SliderInput = forwardRef<HTMLInputElement, SliderInputProps>(
                             </Fragment>
                         )
                     }
+                    {...customInputProps}
                 />
                 {/* eslint-disable react/no-array-index-key */}
                 {steps.length > 0 && !error && (