+++ packages/toast/src/components/base-toast/component.tsx
@@ -3,7 +3,6 @@ import React, {
     type MouseEventHandler,
     type TouchEventHandler,
     useCallback,
-    useEffect,
     useRef,
 } from 'react';
 import mergeRefs from 'react-merge-refs';
@@ -19,7 +18,9 @@ import {
     type ToastPlate as ToastPlateComponent,
     type ToastPlateProps,
 } from '@alfalab/core-components-toast-plate';
-import { useClickOutside, usePrevious } from '@alfalab/hooks';
+import { useClickOutside } from '@alfalab/hooks';
+
+import { useTimer } from './use-timer';
 
 import styles from './index.module.css';
 
@@ -111,25 +112,18 @@ export const BaseToast = forwardRef<HTMLDivElement, BaseToastProps>(
             useAnchorWidth,
             closeWithClickOutside = true,
             fallbackPlacements,
+            client = 'desktop',
             ...restProps
         },
         ref,
     ) => {
         const plateRef = useRef<HTMLDivElement>(null);
-        const timerId = useRef(0);
-        const prevOpen = usePrevious(open);
 
-        const startTimer = useCallback(() => {
-            clearTimeout(timerId.current);
-
-            timerId.current = window.setTimeout(() => {
-                onClose?.();
-            }, autoCloseDelay);
-        }, [autoCloseDelay, onClose]);
-
-        const stopTimer = useCallback(() => {
-            clearTimeout(timerId.current);
-        }, []);
+        const { start: startTimer, stop: stopTimer } = useTimer({
+            open,
+            delay: autoCloseDelay,
+            onTimeout: onClose,
+        });
 
         const handleMouseEnter = useCallback<MouseEventHandler<HTMLDivElement>>(
             (event) => {
@@ -171,22 +165,22 @@ export const BaseToast = forwardRef<HTMLDivElement, BaseToastProps>(
 
         useClickOutside(plateRef, closeWithClickOutside ? handleClickOutside : noop);
 
-        useEffect(() => () => clearTimeout(timerId.current), []);
-
-        useEffect(() => {
-            if (open !== prevOpen && open) {
-                startTimer();
-            }
-        }, [open, prevOpen, startTimer, stopTimer]);
+        const eventHandlers: Record<typeof client, object> = {
+            desktop: {
+                onMouseEnter: handleMouseEnter,
+                onMouseLeave: handleMouseLeave,
+            },
+            mobile: {
+                onTouchStart: handleTouchStart,
+            },
+        };
 
         const props = {
             block,
             titleClassName: cn(titleClassName, styles.title),
             onClose,
-            onMouseEnter: handleMouseEnter,
-            onMouseLeave: handleMouseLeave,
-            onTouchStart: handleTouchStart,
             ref: mergeRefs([ref, plateRef]),
+            ...eventHandlers[client],
         };
 
         if (anchorElement) {