+++ packages/file-upload-item/CHANGELOG.md
@@ -1,5 +1,47 @@
 # @alfalab/core-components-file-upload-item
 
+## 7.2.1
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-icon-button@7.0.3
+
+## 7.2.0
+
+### Minor Changes
+
+<sup><time>13.11.2025</time></sup>
+
+#### [#1921](https://github.com/core-ds/core-components/pull/1921)
+
+- В компонент `FileUploadItem` добавлено свойство `backgroundColor`, позволяющее задать цвет фона для `StatusControl`
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-shared@1.1.1
+    - @alfalab/core-components-typography@5.0.3
+    - @alfalab/core-components-icon-button@7.0.2
+
+## 7.1.1
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-icon-view@4.0.1
+
+## 7.1.0
+
+### Minor Changes
+
+<sup><time>29.09.2025</time></sup>
+
+#### [#1861](https://github.com/core-ds/core-components/pull/1861)
+
+- Реализована анимация прогресса загрузки для режима с неизвестным прогрессом
+- Добавлен новый props `progressBarAvailable` для возможности включения/отключения ProgressBar
+
 ## 7.0.2
 
 ### Patch Changes