+++ packages/select-with-tags/CHANGELOG.md
@@ -1,5 +1,58 @@
 # @alfalab/core-components-select-with-tags
 
+## 9.0.9
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-bottom-sheet@7.4.0
+    - @alfalab/core-components-select@18.2.6
+
+## 9.0.8
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-bottom-sheet@7.3.1
+    - @alfalab/core-components-select@18.2.5
+
+## 9.0.7
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-bottom-sheet@7.3.0
+    - @alfalab/core-components-select@18.2.4
+
+## 9.0.6
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-bottom-sheet@7.2.1
+    - @alfalab/core-components-form-control@13.0.2
+    - @alfalab/core-components-mq@5.0.2
+    - @alfalab/core-components-popover@7.1.1
+    - @alfalab/core-components-select@18.2.3
+    - @alfalab/core-components-tag@9.0.3
+
+## 9.0.5
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-bottom-sheet@7.2.0
+    - @alfalab/core-components-select@18.2.2
+
+## 9.0.4
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-tag@9.0.2
+    - @alfalab/core-components-bottom-sheet@7.1.0
+    - @alfalab/core-components-select@18.2.1
+
 ## 9.0.3
 
 ### Patch Changes